refactor: Separate request and response DTOs - dto 구조의 전반적인 수정 - Request & Response 로 분리
