+++ src/main/java/com/eunsun/travel_mate/dto/request/EmailVerificationRequestDto.java
@@ -1,11 +1,11 @@
-package com.eunsun.travel_mate.dto;
+package com.eunsun.travel_mate.dto.request;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class EmailVerificationDto {
+public class EmailVerificationRequestDto {
 
   private String email;
   private String verificationCode;
